Adds best sales controller.
