+++ backend/src/modules/product/product.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 
-import { ProductController } from './product.controller';
+import { ProductController } from './controllers/product.controller';
+import { SliderController } from './controllers/slider.controller';
 import { ProductSchema } from './product.schema';
 import { ProductService } from './product.service';
 
 @Module({
   imports: [MongooseModule.forFeature([{ name: 'Product', schema: ProductSchema }])],
-  controllers: [ProductController],
+  controllers: [ProductController, SliderController],
   providers: [ProductService],
 })
 export class ProductModule {}
