+++ backend/src/modules/product/product.service.ts
@@ -32,4 +32,14 @@ export class ProductService {
   async getCount(): Promise<number> {
     return this.productModel.count({});
   }
+
+  async findAllSliders(skip: number = 0, limit: number = 3): Promise<Product[]> {
+    return await this.productModel.find(
+      {
+        sliderImage: { $ne: null }
+      },
+      null,
+      { skip, limit }
+    );
+  }
 }
